pyproject.toml
[project]
name = "moai-zero-to-rag"
version = "0.1.0"
description = "Mother-of-AI Phase-1-Zero to RAG"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "uvicorn>=0.34.0",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.13.3",
    "opensearch-py>=3.0.0",
    "requests>=2.32.3",
    "httpx>=0.28.1",
    "docling>=2.43.0",
    "python-dateutil>=2.9.0.post0",
    "sentence-transformers>=5.1.0",
    "gradio>=4.0.0",
    "langfuse>=3.0.0",
    "redis>=6.4.0",
    "python-telegram-bot>=21.0,<22.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.3.0",
]
readme = "README.md"

[dependency-groups]
dev = [
    "anyio[trio]>=4.9.0",
    "asgi-lifespan>=2.1.0",
    "jupyter>=1.1.1",
    "mypy>=1.15.0",
    "notebook>=7.4.4",
    "polyfactory>=2.21.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-aiohttp>=1.1.0",
    "pytest-cov>=6.1.1",
    "pytest-dotenv>=0.5.2",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.5",
    "testcontainers>=4.10.0",
    "types-sqlalchemy>=1.4.53.38",
]
viz = [
    "grandalf>=0.8",
]

[tool.ruff]
line-length = 130
exclude = ["notebooks/**", ".venv/**"]
src = ["src", "tests"]
lint.select = [
  "I",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
env_files = ".env.test"

[tool.mypy]
explicit_package_bases = true
ignore_errors = true


